+++ apps/docs/src/app/core/component-docs/busy-indicator/busy-indicator-header/busy-indicator-header.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-busy-indicator-header',
+    templateUrl: './busy-indicator-header.component.html',
+})
+export class BusyIndicatorHeaderComponent {
+
+}