+++ apps/docs/src/app/core/component-docs/busy-indicator/examples/busy-indicator-toggle-example.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-busy-indicator-toggle-example',
+    templateUrl: './busy-indicator-toggle-example.component.html',
+
+})
+export class BusyIndicatorToggleExampleComponent {
+
+    loading = true;
+
+}