+++ apps/docs/src/app/core/component-docs/popover/examples/popover-simple/popover-example.component.ts
@@ -1,9 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'fd-popover-example',
     templateUrl: './popover-example.component.html',
-    styleUrls: ['popover-example.component.scss']
+    styleUrls: ['popover-example.component.scss'],
+    encapsulation: ViewEncapsulation.None
 
 })
 export class PopoverExampleComponent {