+++ libs/core/src/lib/nested-list/prepared-nested-list/prepared-nested-list.component.ts
@@ -1,7 +1,8 @@
-import { AfterViewInit, ChangeDetectorRef, Component, forwardRef, Input, ViewChild } from '@angular/core';
+import { AfterViewInit, ChangeDetectorRef, Component, forwardRef, Input, QueryList, ViewChild, ViewChildren } from '@angular/core';
 import { NestedListDirective } from '../nested-list/nested-list.directive';
 import { NestedListModel } from '../nested-list-model';
 import { NestedListStateService } from '../nested-list-state.service';
+import { NestedItemDirective } from '../nested-item/nested-item.directive';
 
 /**
  * Component for internal usage, allows to generate the nested list from defined object.
@@ -37,7 +38,21 @@ export class PreparedNestedListComponent implements AfterViewInit {
      * @hidden
      */
     @ViewChild(forwardRef(() => NestedListDirective), { static: false })
-    nestedListDirective: NestedListDirective;
+    _nestedListDirective: NestedListDirective;
+
+    /**
+     * @hidden
+     */
+    @ViewChildren(forwardRef(() => NestedItemDirective))
+    nestedListItems: QueryList<NestedItemDirective>;
+
+    /**
+     * In prepared nested list, nested items should be taken as reference of View, not Content.
+     * There is direct reference to these directives here.
+     */
+    get nestedListDirective(): NestedListDirective {
+        return Object.assign(this._nestedListDirective, { nestedItems: this.nestedListItems });
+    }
 
     /** @hidden */
     constructor (