+++ libs/core/src/lib/utils/decorators/apply-css-style.decorator.ts
@@ -1,0 +1,25 @@
+import { Hash, ELEMENT_REF_EXCEPTION } from '../public_api';
+
+/**
+ * Method decorator to apply css styles to a component through native element
+ * decorator will store original method in variable and wrap it with custom one
+ * component has to implement HasElementRef or CssStyleBuilder interface
+ * more info abour method decorator: https://www.typescriptlang.org/docs/handbook/decorators.html#method-decorators
+ * @param target a component
+ * @param propertyKey name of the method
+ * @param descriptor method
+ */
+export function applyCssStyle(target: any, propertyKey: string, descriptor: PropertyDescriptor): void {
+    const originalMethod = descriptor.value;
+    descriptor.value = function (): Hash<number | string> {
+        if (!this.elementRef) { throw ELEMENT_REF_EXCEPTION; }
+
+        const _styles: Hash<number | string> = originalMethod.apply(this);
+
+        Object.keys(_styles).forEach(key => {
+            (this.elementRef().nativeElement as HTMLElement).style[key] = _styles[key];
+        });
+
+        return _styles;
+    }
+}