+++ libs/core/src/lib/utils/interfaces/css-style-builder.interface.ts
@@ -1,0 +1,5 @@
+import { HasElementRef, Hash } from '../public_api';
+
+export interface CssStyleBuilder extends HasElementRef {
+    buildComponentCssStyle(): Hash<number | string>;
+}